refactor(accounts): replace collect delegate and hooks with events
