+++ backend/main/lib/groupher_server/accounts/events/notify.ex
@@ -1,25 +1,18 @@
-defmodule GroupherServer.Accounts.Delegate.Hooks.Notify do
+defmodule GroupherServer.Accounts.Events.Notify do
   @moduledoc """
-  notify hooks, for upvote, collect, comment, reply
+  notify events, for upvote, collect, comment, reply
   """
   alias GroupherServer.Accounts.Model.User
   alias GroupherServer.Delivery
 
-  # 发布评论是特殊情况，单独处理
   def handle(:follow, %User{} = user, %User{} = from_user) do
-    notify_attrs = %{
-      action: :follow,
-      user_id: user.id
-    }
+    notify_attrs = %{action: :follow, user_id: user.id}
 
     Delivery.send(:notify, notify_attrs, from_user)
   end
 
   def handle(:undo, :follow, %User{} = user, %User{} = from_user) do
-    notify_attrs = %{
-      action: :follow,
-      user_id: user.id
-    }
+    notify_attrs = %{action: :follow, user_id: user.id}
 
     Delivery.revoke(:notify, notify_attrs, from_user)
   end
